Update container IDs to be 64 bits wide

+++ tree_utils.go
@@ -79,7 +79,7 @@ func (self *I3Node) FindFocused() *I3Node {
 }
 
 // Returns a node that has given id.
-func (self *I3Node) FindByID(id int32) *I3Node {
+func (self *I3Node) FindByID(id int64) *I3Node {
 
 	nodes := self.Descendents()
